reorganizes conftest fixtures
